Upd_040824 * Class Track updated - implementation for 'UpdateTracking'. * Class PhysSystem updated - added more methods. * Class Kapow updated - implemented 'ReloadScripts'. * Class AssetManager updated - more findings in asset loader stuff. * DebugMenu lib updated - changed the way menu handler works.

+++ TOD1/QuadTree.cpp
@@ -116,7 +116,7 @@ QuadTreeInfo::QuadTreeInfo()
 const uint8_t QuadTreeNode::SetLodLevel(const Vector4f&, const uint32_t factor)
 {
     const double lodDistanceSqrt = sqrt((double)((uint32_t)LodDistance));
-    const uint32_t lodThresholdAdjusted = lodDistanceSqrt * 64 * LodThreshold;
+    const uint32_t lodThresholdAdjusted = (uint32_t)lodDistanceSqrt * 64 * LodThreshold;
     const uint32_t lodFactorAdjusted = (uint32_t)(float)((float)(factor << 12) * Script::LodFactor);
 
     Lod = 0;
@@ -127,7 +127,7 @@ const uint8_t QuadTreeNode::SetLodLevel(const Vector4f&, const uint32_t factor)
             Lod = 2;
     }
 
-    const uint32_t lodFadeAdjusted = 64 * (FadeThreshold & 127) * (lodDistanceSqrt - Script::MinFadeDist);
+    const uint32_t lodFadeAdjusted = 64 * (FadeThreshold & 127) * ((uint32_t)lodDistanceSqrt - Script::MinFadeDist);
     const uint32_t lodFadeFactorAdjusted = (uint32_t)(float)((float)(factor << 11) * Script::LodFactor);
     const uint32_t lodFadeThresholdAdjusted = lodFadeAdjusted - (lodFadeAdjusted >> (2 - (FadeThreshold >> 7)));
